backend/requirements.txt
pandas
requests
psycopg2-binary
schedule
scikit-learn
plotly
Dash==2.*
dash